Moved&updated localstorage example docs
