+++ v2/emailpassword/advanced-customizations/examples/localstorage/about.mdx
@@ -5,7 +5,8 @@ hide_title: true
 ---
 
 <!-- COPY DOCS -->
-<!-- ./session/advanced-customizations/localstorage-example/about.mdx -->
+<!-- ./session/advanced-customizations/examples/localstorage/about.mdx -->
+
 # Localstorage instead of cookies
 
 Using the features in Advanced customization, we can customize the SDK to use localstorage instead of cookies set on the server. 
@@ -33,3 +34,4 @@ On the frontend we need to do two things:
 We can achieve these by replacing the global fetch function with a "patched" version. This must be done before initializing SuperTokens (calling `SuperTokens.init`) as the SDK uses the fetch function internally.
 
 If you use `axios` (or a similar library) to make requests, you also need to add an interceptor that will do the same.
+<!-- COPY DOCS -->
