+++ v2/session/advanced-customizations/examples/localstorage/about.mdx
@@ -5,6 +5,7 @@ hide_title: true
 ---
 
 <!-- COPY DOCS -->
+<!-- ./session/advanced-customizations/examples/localstorage/about.mdx -->
 
 # Localstorage instead of cookies
 
@@ -12,10 +13,21 @@ Using the features in Advanced customization, we can customize the SDK to use lo
 
 You can find the complete example on [GitHub](https://github.com/supertokens/supertokens-auth-react/tree/master/examples/with-localstorage).
 
-## How the modification works
+## How the modification works123123
 
 The basic idea behind it is that we can use a custom header to transfer tokens. This way, the frontend can access and 
 store them in localstorage or wherever desired and use them to authenticate requests.
+
+:::caution
+Storing tokens in localstorage comes with a few drawbacks:
+
+-   Increased damage potential of XSS attacks because they can now access the longer-term refresh token
+-   Expired tokens are not automatically removed
+-   No session info during server-side rendering
+-   You can't share sessions across subdomains
+
+For a detailed comparison, please see [our blog](https://supertokens.io/blog/cookies-vs-localstorage-for-sessions-everything-you-need-to-know)
+:::
 
 ### Backend
 
@@ -27,7 +39,7 @@ On the backend, we need to do two things:
 
 On the frontend we need to do two things:
 
--   When making a request to the API, we need to add a custom header that contains the tokens we would normally send as cookies
+-   When making a request to the API, we need to add a custom header that contains the tokens we would typically send as cookies
 -   After getting a response from the API, we need to inspect the headers and store the contents of a custom header
 
 We can achieve these by replacing the global fetch function with a "patched" version. This must be done before initializing SuperTokens (calling `SuperTokens.init`) as the SDK uses the fetch function internally.
