+++ v2/session/sidebars.js
@@ -143,10 +143,16 @@ module.exports = {
         },
         {
           type: 'category',
-          label: 'Localstorage example',
+          label: 'Examples',
           items: [
-            "advanced-customizations/localstorage-example/about",
-            "advanced-customizations/localstorage-example/guide"
+            {
+              type: 'category',
+              label: 'Using localstorage instead of cookies',
+              items: [
+                "advanced-customizations/examples/localstorage/about",
+                "advanced-customizations/examples/localstorage/guide"
+              ],
+            },
           ],
         },
       ],
